Fixing more of my past shortcuts
